Massive rework of the code.  Current should serve as a much better base to explore the larger numbers of models and their parameters.
Initial models (test set) are working, as is export and some R based plotting functions.  the plan if for all statistical analysis and plotting to happened in R.  The stats function is now deprecated, but will be left in place as it still offers some convenience for testing and development.

The read me file is not totally dated and need to be reworked.

The test file should also become what is says it is.  ….Tests for each of models and the main methods for Trials() need to be put in place.